Issue #327: SQL Source Module Renaming
