+++ metricshub-linux/pom.xml
@@ -78,7 +78,7 @@
 		</dependency>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
-			<artifactId>metricshub-sql-h2-source-extension</artifactId>
+			<artifactId>metricshub-sql-source-extension</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 	</dependencies>
