--- metricshub-sql-extension/src/main/java/lombok.config
@@ -1,2 +1,0 @@
-lombok.nonNull.exceptionType = IllegalArgumentException
-lombok.addLombokGeneratedAnnotation = true
