+++ pom.xml
@@ -116,7 +116,7 @@
 		<module>metricshub-wmi-extension</module>
 		<module>metricshub-winrm-extension</module>
 		<module>metricshub-ping-extension</module>
-		<module>metricshub-sql-extension</module>
+		<module>metricshub-sql-source-extension</module>
 		<module>metricshub-hardware</module>
 		<module>metricshub-agent</module>
 		<module>metricshub-doc</module>
